Adopt SF-0023 ProgressManager
Progress now flows down into SwiftNASR as a `Subprogress` minted from the
caller's own `ProgressManager`, rather than up out of it as a `Progress`
handed to a handler block. Six public signatures trade their
`withProgress:` handler for `progress: consuming Subprogress? = nil`:
`NASR.load`, `NASR.parse`, `Loader.load`, `Downloader.load`,
`Distribution.readFile`, and `Distribution.readFileRaw`.

The platform floor rises to macOS/iOS/tvOS/watchOS/visionOS 27 and the
manifest to tools-version 6.4, which is where `ProgressManager` starts
existing: it carries no lower availability annotation on Apple platforms,
it is `@available(FoundationPreview 6.4)` on Linux, and `MacOSVersion.v27`
is itself `@available(_PackageDescription 6.4)`. CI follows: the Darwin
legs collapse into one `xcode-27` job driven by `maxim-lobanov/setup-xcode`
(no earlier image can supply the macOS 27 SDK), the Linux 6.3 leg is
dropped, and `.github/actionlint.yaml` declares `xcode-27` so
`actionlint` keeps checking runner labels.

`DirectoryDistribution` double-counted every file it read, adding each
line's length on top of the length of the chunk the line arrived in and
reporting 35 bytes read from a 21-byte file. `Progress` clamped
`fractionCompleted` to 1.0 and hid it; `ProgressManager` reports the
overshoot. The file's byte accounting is now pinned through the new
`completedByteCount`/`totalByteCount` metadata, which every read and
download reports alongside its unit counts.

Also: `CSVParser` sheds its vestigial `progress` and `bytesRead`
requirements, a diagnosed record no longer stalls the fixed-width parse
progress short of its total, and the E2E progress bar follows
`Observations.untilFinished` over the observable root manager instead of
re-reading `fractionCompleted` on a 10 Hz timer.

+++ Tests/SwiftNASR_E2E/ProgressWeights.swift
@@ -1,10 +1,10 @@
 import SwiftNASR
 
 /// Weight for the initial loading phase (relative to parsing weights).
-let loadingWeight: Int64 = 10
+let loadingWeight: Int = 10
 
 /// Returns the weight for a record type based on format.
-func weight(for recordType: RecordType, format: DataFormat) -> Int64 {
+func weight(for recordType: RecordType, format: DataFormat) -> Int {
   recordTypeRegistry[recordType]?.weight(for: format) ?? 1
 }
 
@@ -14,7 +14,7 @@ func availableRecordTypes(for format: DataFormat) -> Set<RecordType> {
 }
 
 /// Calculates the total progress weight for selected record types.
-func totalWeight(format: DataFormat, selectedRecordTypes: Set<RecordType>) -> Int64 {
+func totalWeight(format: DataFormat, selectedRecordTypes: Set<RecordType>) -> Int {
   let recordTotal = selectedRecordTypes.reduce(0) { $0 + weight(for: $1, format: format) }
   return loadingWeight + recordTotal
 }
